Update engine.js
Redid the Loader/Resource classes to handle loading more efficiently. The Loader no longer has a loop that checks if its resources are loaded yet or not. Instead, Resources bind their functions to the onload/onerror events.

Redid how the Player handles resizing. It now uses CSS scaling, as the MDN docs note that this is more efficient. This code still needs to be cleaned up.

Began working on Tile and TileMap classes.