Update engine.js
Updated keyboard input handler for better use with the new TextBox gameObject.
Added some common keys to the Input.keys

Updated Font to use some constants in place of 'centered', 'left', 'bottom', etc.

Fixed bug in TextRenderer that caused strings starting with an in-line statement ({Font.BOLD}, for example) to not work properly.

Allowed for the use of multiple Font modifiers in a single in-line statement ({Font.BOLD;Color.RED}, for example).

Fixed bug where the TextRenderer was using mis-named variables.

Added prototype of a TextBox object. Currently allows for typing and moving the cursor.