Update engine.js
Added a safeEval method to the Util class. This is used for the TextRenderer when evaluating in-line styling.

Added some documentation for showing how to use in-line styling in the TextRenderer class.

Added maxLength attribute to the TextBox class.
Added the ability to hold down keys and have them move/delete/place a character with increasing speed. Control this speed through the timeHoldPerChar and timeHoldMultRate members.