Update engine.js
Fixes to destroying components.

Removed duplicate code which would try to remove an InputComponent twice.

Fixed GameObject.destroyComponents not working due to modifying the array while iterating through it.

Removed un-needed code when destroying a RigidBody

Fixed method that fixes Collider collision list -- was checking for this.gameObject to be undefined when it should have checked for it to be null.

Removed all Component destroy methods, due to duplicate calls being made. Component.remove is called when the component is destroyed (or its gameObject is destroyed), so all of the extra logic has been moved to the Component.remove methods.
This means that in the future new classes that extend Component will only need to override the remove() method to ensure their removal/deletion is handled properly.

Fixed Collider.onStay never being called.