Update engine.js
Changed GameObject.addComponent to GameObject.attach, for clarity reasons.
Changed Component.addTo to Component.attachTo, to match the change to GameObject.

Added Component.onAttach method which can be overwritten and is called when the component is attached to a GameObject. This is mostly to fix a bug in Collider where the bound was not being set properly if negated on creation.

Separated some methods from the Util class to a new class called Random. These methods include: Util.randRange, Util.randStr, and Util.arrSample. These have been renamed to Random.range, Random.string, and Random.sample.